authorization added

+++ routers/posts.py
@@ -7,21 +7,21 @@ from typing import Annotated
 import models
 from database import Base, engine, get_db
 from sqlalchemy.ext.asyncio import AsyncSession
-
+from auth import CurrentUser
 
 router = APIRouter()
 
 #get post
 @router.get("", response_model=list[PostResponse])
-async def get_posts(db: Annotated[AsyncSession, Depends(get_db)]):
+async def get_posts(db: Annotated[AsyncSession, Depends(get_db)], current_user: CurrentUser):
     result = await db.execute(select(models.Post).options(selectinload(models.Post.author)))
     posts = result.scalars().all()
     return posts 
 
 #create post
 @router.post("/create", response_model=PostResponse)
-async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
-    result = await db.execute(select(models.User).where(models.User.id == post.user_id))
+async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_db)], current_user: CurrentUser):
+    result = await db.execute(select(models.User).where(models.User.id == current_user.id))
     user = result.scalars().first()
 
     if not user:
@@ -30,7 +30,7 @@ async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_
     new_post = models.Post(
         title = post.title,
         content = post.content,
-        user_id = post.user_id
+        user_id = current_user.id
     )
     db.add(new_post)
     await db.commit()
@@ -39,20 +39,25 @@ async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_
 
 
 @router.get("/{post_id}", response_model=PostResponse)
-async def get_post(post_id:int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def get_post(post_id:int, db: Annotated[AsyncSession, Depends(get_db)], current_user: CurrentUser):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id).options(selectinload(models.Post.author)))
     post = result.scalars().first()
     return post
 
 #update post partial
 @router.patch("/{post_id}", response_model=PostResponse)
-async def update_post_partial(post_id: int, post_data: postUpdateParitial, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_partial(post_id: int, post_data: postUpdateParitial, db: Annotated[AsyncSession, Depends(get_db)], current_user: CurrentUser):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(
             status_code=404,
             detail="Post not found"
+        )
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=403,
+            detail="You are not authorized to update this post"
         )
     update_data = post_data.model_dump(exclude_unset=True)
     for key, value in update_data.items():
@@ -64,7 +69,7 @@ async def update_post_partial(post_id: int, post_data: postUpdateParitial, db: A
 
 #update post fully
 @router.put("/{post_id}", response_model=PostResponse)
-async def update_post_fully(post_id: int, post_data: PostUpdate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_fully(post_id: int, post_data: PostUpdate, db: Annotated[AsyncSession, Depends(get_db)], current_user: CurrentUser):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
     if not post:
@@ -72,6 +77,12 @@ async def update_post_fully(post_id: int, post_data: PostUpdate, db: Annotated[A
             status_code=404,
             detail="Post not found"
         )
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=403,
+            detail="You are not authorized to update this post"
+        )
+
     update_data = post_data.model_dump()
     for key, value in update_data.items():
         setattr(post, key, value)
@@ -82,13 +93,18 @@ async def update_post_fully(post_id: int, post_data: PostUpdate, db: Annotated[A
 
 #delete post
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def post_delete(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def post_delete(post_id: int, db: Annotated[AsyncSession, Depends(get_db)], current_user: CurrentUser):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(
             status_code=404,
             detail="Post not found"
+        )
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=403,
+            detail="You are not authorized to delete this post"
         )
     await db.delete(post)
     await db.commit()
